Replace `Byte` and `Short` with `Int`

+++ src/commonMain/kotlin/response/AvailableDataCenter.kt
@@ -12,5 +12,5 @@ import kotlinx.serialization.Serializable
 @JsExport @Serializable class AvailableDataCenter(
     val name: String,
     val region: String,
-    val worlds: ShortArray,
+    val worlds: IntArray,
 )
