+++ src/commonMain/kotlin/response/RecentlyUpdatedItem.kt
@@ -5,8 +5,8 @@ import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
 @JsExport @Serializable class RecentlyUpdatedItem(
-    @SerialName("itemID") val itemId: Short,
+    @SerialName("itemID") val itemId: Int,
     val lastUploadTime: Long,
-    @SerialName("worldID") val worldId: Short,
+    @SerialName("worldID") val worldId: Int,
     val worldName: String,
 )
